lesson 8: change Double to BigDecimal in currencies dto and now client put currencies into map by CharCode, instead of list

+++ currency-app/src/main/java/ru/tbank/currencyapp/dto/CurrencyConvertedDTO.java
@@ -1,8 +1,10 @@
 package ru.tbank.currencyapp.dto;
 
+import java.math.BigDecimal;
+
 public record CurrencyConvertedDTO(
         String fromCurrency,
         String toCurrency,
-        Double convertedAmount
+        BigDecimal convertedAmount
 ) {
 }
