+++ currency-app/src/main/java/ru/tbank/currencyapp/dto/CurrencyRateDTO.java
@@ -1,7 +1,9 @@
 package ru.tbank.currencyapp.dto;
 
+import java.math.BigDecimal;
+
 public record CurrencyRateDTO(
         String currency,
-        Double rate
+        BigDecimal rate
 ) {
 }
