+++ currency-app/src/main/java/ru/tbank/currencyapp/dto/cb/CBCurrencyResponseDTO.java
@@ -1,12 +1,14 @@
 package ru.tbank.currencyapp.dto.cb;
 
+import java.math.BigDecimal;
+
 public record CBCurrencyResponseDTO(
         Integer numCode,
         String charCode,
         Integer nominal,
         String name,
-        Double value,
-        Double vUnitRate
+        BigDecimal value,
+        BigDecimal vUnitRate
 ) {
     public static CBCurrencyResponseDTO fromValute(ValuteDTO valuteDTO) {
         return new CBCurrencyResponseDTO(
